Augmentation to the type system to support user defined conversion classes and corresponding changes to the JNI type system to correct issues including locking problem

+++ native/python/include/jp_pythontypes.h
@@ -361,7 +361,7 @@ public:
 		return contents.size();
 	}
 
-	PyObject* operator[](int i)
+	PyObject* operator[](ssize_t i)
 	{
 		return contents[i].get();
 	}
